requirements.txt
gradio
langchain
langchain-openai
langchain-community
langchainhub==0.1.15
python-dotenv
openai
pypdf

